Update to the About section
Adds content to the about, removes the redundancy in the index.css.

//TODO:
Fix the “About” tab in the navbar so it actually takes you to the about
page
Fix the text at the top? It is really hard to read
Get the information from Mrs. Bis
Is there any place that has a more in-depth description of the main
idea of our club?